Trait fiat_filter_rpc_runtime_api::FiatFeeRuntimeApi

pub trait FiatFeeRuntimeApi<Block: BlockT, Balance>: Core<Block> where
    Balance: Codec + MaybeDisplay, {
    fn get_call_fee_dock(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        uxt: Block::Extrinsic
    ) -> Result<Result<Balance, Error>, ApiError> { ... }

    fn get_call_fee_dock_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        uxt: Block::Extrinsic
    ) -> Result<Result<Balance, Error>, ApiError> { ... }
}

Provided methods

fn get_call_fee_dock(
&self,
__runtime_api_at_param__: &BlockId<Block>,
uxt: Block::Extrinsic
) -> Result<Result<Balance, Error>, ApiError>

Accepts any extrinsic, signature not needed. Returns the call fee fee in µDOCK as Balance (Balance is u64 in current runtime) If the extrinsic is not one of the calls supported by FiatFilter, returns an error

fn get_call_fee_dock_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
uxt: Block::Extrinsic
) -> Result<Result<Balance, Error>, ApiError>

Accepts any extrinsic, signature not needed. Returns the call fee fee in µDOCK as Balance (Balance is u64 in current runtime) If the extrinsic is not one of the calls supported by FiatFilter, returns an error